requirements.txt
# AradhyeMusic — Python dependencies (Python 3.12 compatible)

# Telegram core
pyrofork>=2.3.34
kurigram>=2.2.6
tgcrypto>=1.2.5
py-tgcalls>=2.1.5
pykeyboard>=0.1.7

# Async / HTTP
aiofiles>=24.1.0
aiohttp>=3.10.10
httpx>=0.27.2
requests>=2.32.3
asyncio

# Database
motor>=3.6.0
dnspython>=2.7.0

# Media / audio
ffmpeg-python>=0.2.0
hachoir>=3.3.0
gTTS>=2.5.4
pillow>=11.0.0
opencv-python-headless>=4.10.0.84
numpy>=2.1.3

# Music sources
yt-dlp>=2024.10.22
youtube-search>=2.1.2
youtube-search-python>=1.6.6
py-yt-search>=2.0.0
spotipy>=2.24.0

# Utility
beautifulsoup4>=4.12.3
gitpython>=3.1.43
heroku3>=5.2.1
psutil>=6.1.0
pytz>=2024.2
python-dotenv>=1.0.1
pyyaml>=6.0.2
speedtest-cli>=2.1.3
telegraph>=2.2.0
unidecode>=1.3.8
